run-installer.sh: Updated to use Hax2.app directly
Updated run-installer.sh so that you can copy Hax2.app directly onto
the USB stick, and it'll know that Hax2Lib.dylib is embedded therein.
Hopefully this will make the patcher a little easier to use.

(In the process, I fixed a couple of other incredibly minor issues that
basically aren't worth mentioning in detail.)