Split some code from run-installer.sh to insert-dylib.sh
This begins a refactoring by taking the library insertion code from
run-installer.sh and copying it to insert-dylib.sh. (The pmset code for
disabling display and system sleep comes along for the ride.)

run-installer.sh still contains the original code and does not yet call
insert-dylib.sh, so there is not yet any potential for behavioral change.
In the next phase, run-installer.sh will call insert-dylib.sh instead of
containing a copy of the same code. Once that passes testing, I'll probably
merge that from the dev-v0.0.4 branch to the dev-v0.0.3 branch, to minimize
differences in code between 0.0.3 and 0.0.4. (And after that, I'll bump
the version number on the 0.0.4 branch.)

On the dev-v0.0.4 branch, the phase after that will be to rearrange things
so that insert-dylib.sh is called not from run-installer.sh but from
set-vars.sh, as part of a larger change in how the patcher works.