Finish up the v0.0.4 release
I've updated the README to correspond with this release's new behavior,
and I've bumped the version number to 0.0.4.

At this point, once you have the patched USB in hand, the process of
doing the basic installation is really the same as on a supported Mac,
except that you have to open Terminal, run one shell script, and close
Terminal again before running the Installer.