Add extracted Hax2Lib.dylib to payloads
I figure it's less risky to just add the one dylib to the repo than a
full app bundle. It's all that my patcher actually needs, anyway. With
the original zip files bundled as well, we're covered for any future
needs or anyone else who needs source (which reminds me, I need to update
the original-zips readme as well).