Make patcher & unpatcher sync when done
Both the patcher and unpatcher are causing enough disk activity on the
USB stick that it seems advantageous to have both of them sync when
they're done. (In my testing, it seems to cause my USB drive's light
to stop blinking sooner, which is a nice bonus.)

I can remove this later if anyone tries this patcher on a genuinely busy
system and sync gets stuck, but in the meantime I'll include this because
I doubt it'll happen and I think syncing at the end is a good idea.