Add WiFi patch script, finally!
This script must run from the installer USB, and it does not yet handle
mounted snapshots (rebooting from the installer USB has always been enough
for me to see the actual volume rather than the snapshot get mounted, in
my testing so far). However, it's a start, and in my testing it works so far!