Add unpatch-kexts.sh (does what it says)
This script basically tries to undo whatever modifications patch-kexts.sh did.
(There is the possibility that it may not work reliably if kexts are patched
or uninstalled using means other than patch-kexts.sh.)

Since it's a separate shell script, and it has error-checking similar to
patch-kexts.sh, I see little risk in adding it to the v0.0.6 release.

Eventually I may merge it back into patch-kexts.sh as an --uninstall
option or something like that, but that's definitely too risky for v0.0.6,
and I would need to think about it further to really know whether that merge
would be a good or bad idea.