v0.0.8: Try harder to find payloads directory
If micropatcher.sh fails to find the payloads directory, it's probably
because the PWD isn't the directory that contains micropatcher.sh. So,
it will now try to figure out its own directory, change to that, and
check again for the payloads directory.

This should fix a patcher failure that is happening in the real world.
That justifies releasing v0.0.8 with just this fix.