patch-kexts.sh: Add --hda option again
Add --hda option to install 802.11n and AppleHDA kexts, but not HD3000
kexts. I added this option because there are some easily reproducible
kernel panics when using the HD3000 kexts. (On the other hand, workarounds
for the kernel panics seem possible, and performance is definitely much
better on a 13" 2011 MBP with the HD3000 kexts, even though it's not
truly accelerated.)

(While I was at it, I also fixed some progress output which I forgot to
update in previous commits.)