Update AppleHDA to latest High Sierra version
It turns out that AppleHDA was updated at some point after the High
Sierra 10.13.6 release, so I've updated it to the one from the latest
High Sierra security update (build 17G13035). I took the opportunity to
rename the AppleHDA kext zip file in line with the others in this patcher
and to clean up the code in patch-kexts.sh a little as well.