(un)patch-kexts.sh: use kmutil create instead of kmutil install
In a MacRumors forum post, testheit suggested the --bundle-path parameter
of kmutil install as a possible way of getting LegacyUSBInjector to work.
(As it turns out, that does seem to work, at least within certain
constraints.)

However, for the sake of being able to troubleshoot this later in the event
that it doesn't work right for some reason, I'm separating the
kmutil install -> kmutil create transition into its own commit. (In fact,
if I need to go back and make a v0.0.12 release later, this commit is
probably the entirety of the code that should change from v0.0.11 to v0.0.12,
more or less.)