patch-kexts.sh: Add LegacyUSBInjector
This commit kind of cheats by bundling it in with the other --2011
kexts, but it turns out that LegacyUSBInjector would be good to install
on Sandy Bridge MacBook Pros anyway (and dosdude's Catalina Patcher does
this, in fact). But at least it gets the initial code in there, and another
commit can clean things up.

Speaking of which, an upcoming commit will add telemetry plugin
deactivation, which will allow actual testing on my 2009 white MacBook.