Substantial (but incomplete) README update
This updates lots of stuff in the README, but it's an incomplete update. If this was going straight to the main branch, this would be unacceptable quality for a README, but there's no problem in committing this to the dev branch -- forthcoming commits will complete and fix the README of course.