Add BCM5701 Ethernet kext from 10.15.6
This is needed for fixing Ethernet problems from beta 3 (and to a lesser
extent, beta 1 and 2 as well).

A previous commit already updated unpatch-kexts.sh for removing this
kext modification. I still need to update patch-kexts.sh for installing it.
(I plan to whitelist based on build numbers, because the fact that it's
also affecting Thunderbolt Gigabit Ethernet adapters makes me think it
might be a bug that Apple will fix at some point rather than an
intended compatibility break.)