patch-kexts: BCM5701 kext install (and clean up output)
Changes to install the BCM5701 kext from Catalina 10.15.6, if the
Big Sur build number matches a whitelist. (The whitelist is because
I'm not sure if this is a bug that Apple might fix later in Big Sur
development.)

This commit also has some other changes to clean up the output of the
script a bit. Ideally all this wouldn't be mixed into one commit, but
I ended up doing it as part of debugging dumb typo bugs while adding the
build whitelist checking, etc.