patch-kexts: Move original BCM5701 to correct location!
Oh man, this one's bad. When the new BCM5701 was being installed, the
original was being moved to the wrong location -- and probably not the
location that you would first guess based on looking at the source code!

(It was getting moved into a subdirectory of AppleHDA.kext.original.)

This happened because of a stupid copy-and-paste error. (I'll probably
restructure this code as part of v0.0.18.)