patch-kexts.sh: Initial fixes to handling of mounted snapshots
Two bug fixes:

(a) One of the are-we-in-recovery-environment if-statements didn't work
    because of a typo.

(b) After mounting the underlying volume, trying to remount that mount as
    read-write when it's already read-write causes errors under Big Sur
    beta 3, so don't do that. (I swear this worked under beta 1...)

There are some outdated comments that need to be fixed, and possibly some
other code that needs fixing to make things more robust, but it's working
under a live Big Sur installation that's booted from a snapshot! I also
need to retest under the recovery environment.