patch-kexts: List temp mntpt; try to unmount at end
If patch-kexts had to create and use a temporary mountpoint to mount the
underlying volume, it now tells the user the path to the temporary
mountpoint, and it attempts to unmount it once it's done with everything
else.

Without attempting to unmount, running this script a second time without
rebooting is *guaranteed* to fail.