kc-list-kext-paths: Fix Xcode compatibility
Previously this script didn't work with Xcode 10.1 under High Sierra.
There are probably other versions of Xcode that didn't work either, but
it's a moot point now because, with this commit, it should work with all
Xcode versions.

(Add a comment to document things a little better while I'm at it.)