fork HaxLib from Hax3
This adds a fork of the HaxLib from Hax3. As of this commit, it only
adds comments at the top of the file. However, I decided it would be
best to do an initial commit with no code changes. That way, it will
be easy to see all of the changes that I make by looking at the
source control history.