add payloads/remount-sysvol.sh
This script remounts the system volume read-write, or mounts the
underlying volume if the system is running from a snapshot.

This is an initial version. I expect to add some minor changes before
the v0.0.21 release. Another script, rebuild-kc.sh, is also on the way,
as soon as I test it a little more.