HD3000: Remove VA bundles to fix kernel panic
On Macs with Intel HD3000 graphics, there were some kernel panics.
It looked possibly video acceleration-related to me, so I removed
the *VA* parts of the HD3000 driver (I don't think video acceleration
is working anyway), and I'm not having any more kernel panics. The
HD3000 framebuffer also seems to still be otherwise working.

Even if it turns out that this is not a 100% fix, this probably solves
the single biggest cause of HD3000 kernel panics in Big Sur.