micropatcher.sh: TCC check for Catalina
This does a check for necessary TCC permissions on Catalina. If it fails,
it resets the TCC permissions for Terminal and tries one more time.

This should fix various issues with micropatcher.sh failing to apply
the patch properly on Catalina.

Eventually unpatch.sh needs similar fixes, but all the real world bug
reports have involved micropatcher.sh, so that can wait for another
release.

There may be some upcoming commits to make this a little prettier and
more user friendly.