README: Document remount-sysvol.sh
While I was writing this, I noticed a bug in remount-sysvol.sh. For now I wrote these instructions to account for the bug, but I plan to actually fix the bug in the next patcher release. (The rest of the changes in this release are important enough that I'm not going to hold it up just to fix this one bug.)