Fixed Changes to allow mount -uw / on Big Sur
This is unnecessary. 
When this snapshot is created, It cannot be deleted from a recovery terminal with 'diskutil apfs deleteSnapshot' as this blessing becomes un-purgeable. 
This breaks the trick that makes 'mount -uw /' work. (If the machine has no apfs snapshots on their root partition and has authenticated root disabled. 
I'm considering this pull request as a vital update as many developers and hackintosh users depend on this trick for development.