setvars.c: Make bootArgs[] static
This change makes sure the bootArgs[] string is stored contiguously within
the binary that's generated by gcc. (Seriously, it's *not* stored
contiguously without this change!)

This might be good in case anyone needs to binary patch bootArgs[] later.
Even if that is never needed, it doesn't seem to do any harm.