(un)patch-kexts.sh: Add --setBoot to bless
At least on my MacBookPro8,1, creating the snapshot with bless causes
the startup disk selection to get lost, at least with beta 6 (I didn't
test with beta 5). This makes startup take much longer than it should,
so use --setBoot to avoid that.

This fix is needed for stability, not just performance. There are
plenty of situations where, in the absence of startup disk selection
information, the EFI firmware might choose to boot the wrong partition.

(This should work OK with previous betas as well.)