micropatcher.sh: Store installer app name in variable
This is going to be needed for installing the trampoline patch.

However, at this point, this code is starting to look a bit ugly to
me, and I'd like to fix it to handle the golden master (since that
might be coming soon). So I think I'll clean up this code before I
continue.