install-setvars.sh: initial commit
This script installs the setvars EFI program onto the EFI partition of
the patched installer USB. (The installer USB must be formatted with a
GUID partition table.) Thanks to Travis Parker for the idea. This is
part of the resolution for issue #24.

I still need to do some more testing of the error handling, and I
might rename it if I can think of a better name in the next day or so,
but so far it seems to work.