trampoline: turn off display/system sleep
Since we're installing on unsupported Macs, and there's a good chance
(especially on pre-2012 Macs) that the installer USB is missing the
necesssary framebuffer kexts for proper power management, the trampoline
should disable display sleep and system sleep.

(On a 2012 or newer Mac, there should be no harm from this change, either.
It's really the same as what previous versions of the patcher did, but
the trampoline is now used instead of the old scripts.)