(un)patch-kexts: Remove --bootefi option from bless
It turns out this option does stuff under /System/Library/CoreServices
that guarantees failure of delta update filesystem verification. In
my testing, it also appears that this option isn't necessary, at
least not on beta 6.

This change alone is not sufficient to fix unpatch-kexts so that
filesystem verification passes afterward, but it's necessary.

(If it turns out that some earlier version of Big Sur requires this,
then I can add extra checks for that later on.)