Add a copy of the previous highvoltage12v WiFi patch
This will make it easier for people to switch back to the old
version of the WiFi patch, if they want to do that.

(Ideally this would be accompanied by a command line option to
use this patch in patch-kexts.sh. I might first try doing a
different kind of WiFi patch and see how that goes, though.)