bring run-installer.sh back from the dead
ASentientBot has a method of doing clean installs of (at least some)
Big Sur betas which don't have full installers:

https://forums.macrumors.com/threads/macos-11-big-sur-on-unsupported-macs-thread.2242172/page-173?post=28930995#post-28930995

There *may* be a way to use this script to make this method easier.
I'm not sure yet. But making this script work again is the first step
in finding out.

(I have not tested it yet, so these changes may not be sufficient,
but they're basically necessary.)