Update README; bump version to v0.3.1
As of this writing, I'm actually not done testing it with beta 9 yet,
but it has gotten far enough in the installation process that I would
be very surprised if it fails.

I'll wait until testing is done -- I think it should take less than
an hour -- then, if it succeeds, I'll merge this into the main branch
and release it.