setvars: Update source for beta 9 SIP
As of beta 9 (although maybe the change actually happened in beta 7,
I'll have to check later), "csrutil disable" now sets the first byte
to 0x7f, not 0x77. So, to properly disable SIP, setvars needs to
adopt the same change.

This change just changes the source. The binaries will be updated in
another commit.