README: Try to make step 15 easier
Thanks to thatsmeonlyme on the MacRumors Forums for suggesting that
this step needed improvement. I decided to improve it a different way,
but let's see how this works out. (Eventually I need to make patch-kexts.sh
easier to use, but there are other things I want to fix first, so hopefully
this will be a sufficient stopgap measure.)