add kmutil binaries from beta 8
kmutil in beta 9 shows far less informative error messages. Big Sur
beta 8's kmutil *will* be needed in the future for troubleshooting
purposes. I still need to modify (un)patch-kexts.sh to make use of this,
and that's probably not happening in this release, but it's a certainty
that these will be needed so I may as well add them now. (Well, maybe
only one is actually necessary, but I'll include both, for now at least.)