Place the kmutil files on the USB
This modifies micropatcher.sh to copy the kmutil binaries onto the USB
and modifies unpatch.sh to remove them.

I'm not sure if I'll add any options to (un)patch-kexts.sh now, but
at least I might be able to give custom instructions on the MacRumors
forum thread or something if needed, for the moment.