remount-sysvol.sh: fix issue #59 here too
Basically a copy-and-paste of the fix from patch-kexts.sh.

By the way, the original idea with remount-sysvol.sh was to make it
a component that could be reused by patch-kexts.sh, but I had trouble
making that work reliably while still maintaining compatibility with
the ancient version of bash that Apple ships. Switching to zsh would
be the obvious solution, except Apple doesn't ship zsh on the installer
USB's BaseSystem.

Just copying code is probably the best solution for now, until keeping
patch-kexts.sh and remount-sysvol.sh in sync becomes a bigger problem,
or until all the other bigger problems are solved first and I actually
have a chance to revisit the issue properly.