bump version to 0.4.1; update testing info
Actually, the testing info isn't quite correct -- in addition to
beta 9 testing, I also did some limited testing with beta 3, 4, and 6.
However, I'm trying to get this out before the big Apple event, and
the README is already a bit verbose as it is, so I think it's accurate
enough.