patch-kexts: Fix USB detection for GM
The USB stick detection was poised to fail once GM happens. (I fixed
this for at least some of the other scripts in v0.2.0.)

There's probably stuff about this code that I should change, but I need
to finish this release quickly to fix the beta 10 issue, so this will do
for now.