unpatch-kexts: Use --bootefi and current bless
Use the --bootefi option instead of using an old version of bless.
As far as I can tell right now, beta 10's bless fixes the bug that
required omitting --bootefi in the first place.

If it turns out that I'm mistaken and that the old bless is needed
after all, I'll just revert this change.