create and use a "bin" directory
I expect that the number of binaries is going to go down soon (I
am hoping to eventually remove the beta 9 bless and beta 3 kmutil),
but having `patch.19h2` on the root directory of the patched USB
stick is going to make tab completion work less effectively, so
I think adding a bin directory is a good idea anyway.

This commit is going to cause some breakage, but I'll fix it with
upcoming commits.