run-installer.sh: add more code (but commented out)
As an experiment, I tried changing the final line, that actually runs
the installer, so that all error/diagnostic messages from the installer
binary are redirected to /dev/null.

This commit adds that code but also comments it out. I'm not sure
which way is best, but I expect approximately nobody (other than myself)
will use this script anyway. (I may use this script as the basis for
something more interesting, however.)